requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
fastapi-cors
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart